# swagger.yml
openapi: 3.1.1
info:
    title: shopez e-commerce API
    version: 1.0.0
    description: A REST API for working with the shopez e-commerce platform

paths:
    /auth/register:
        post:
            summary: Registers a new user
            description: Registers a new user
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user'
            responses:
                201:
                    description: Created
                400:
                    description: Bad Request
                409:
                    description: Conflict
    /auth/login:
        post:
            summary: Logs in an existing user
            description: Logs in an existing user
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user'
            responses:
                200:
                    description: OK
                400:
                    description: Bad Request
                401:
                    description: Unauthorised
    /auth/logout:
        post:
            summary: Logs out a currently logged in user
            description: Logs out a currently logged in user
            responses:
                200:
                    description: OK
    /customers/{customerId}:
        get:
            summary: Get customer information
            description: Get customer information
            responses:
                200:
                    description: OK
                500:
                    description: Internal Server Error
        put:
            summary: Update customer information
            description: Update customer information
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/customer'
            responses:
                200:
                    description: OK
                400:
                    description: Invalid Argument Provided
    /products:
        get:
            summary: Gets information of all products
            description: Gets information of all products
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product'
        post:
            summary: Adds a new product to the system
            description: Adds a new product to the system
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/product'
            responses:
                200:
                    description: OK
                400:
                    description: Bad request
                409:
                    description: An unexpected error occurred and the product record was not created
    /products/{productId}:
        get:
            summary: Gets information of a specified product
            description: Gets information of a specified product
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/product'
                400:
                    description: Item not found
        put:
            summary: Updates a single product in the system
            description: Updates a single product in the system
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/product'
            responses:
                200:
                    description: Product record {productId} updated
                400:
                    description: Bad request
                409:
                    description: An unexpected error occurred and the product record was not updated
        delete:
            summary: deletes a single product from the system
            description: deletes a single product from the system
            responses:
                200:
                    description: Product record {productId} deleted
                404:
                    description: Item not found
components:
    schemas:
        customer:
            type: object
            properties:
                firstName:
                    type: string
                lastName:
                    type: string
                email:
                    type: string
                phone:
                    type: string
            required: [email]
        product:
            type: object
            properties:
                description:
                    type: string
                price:
                    type: integer
                    description: Price in cents
                brand:
                    type: string
                model:
                    type: string
        order:
            type: object
            properties:
                date:
                    type: date
                status:
                    type: integer
        user:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                    format: password
            required: [username, password]